v0.0.0-drafts
Not sure this will be of any help to you; but here's the current status of what I was working on; I may have messed up a lot of things you had right, I'm not sure. The issue I'm currently having is the submit button for logging in is not triggering the callback.